requirements.txt
numpy>=1.21.0
torch>=1.9.0
librosa>=0.9.0
gradio>=3.50.0
python-dotenv>=0.19.0
moviepy>=1.0.3
opencv-python>=4.5.3
transformers>=4.11.0
soundfile>=0.10.3
scipy>=1.7.1
pillow>=8.3.2
tqdm>=4.62.3
requests>=2.26.0
pydub>=0.25.1
ffmpeg-python>=0.2.0
huggingface-hub>=0.4.0 